job-service/requirements.txt
fastapi==0.104.0
uvicorn==0.23.2
sqlalchemy==2.0.23
alembic==1.12.1
asyncpg==0.29.0
redis==5.0.1
pydantic==2.5.1
pydantic-settings==2.1.0
python-multipart==0.0.6
prometheus-client==0.19.0
kubernetes==28.1.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1
opentelemetry-api==1.20.0
opentelemetry-sdk==1.20.0
opentelemetry-exporter-otlp==1.20.0
opentelemetry-instrumentation-fastapi==0.41b0
